add items page, move map to component
