+++ src/pages/Items.js
@@ -1,0 +1,12 @@
+import Map from "../components/Map";
+import Page from "../components/Page";
+
+export default function ItemsPage () {
+
+
+  return (
+    <Page>
+      <Map />
+    </Page>
+  )
+}
